rename entity cache to navigation provider
