+++ libs/lib-editing/src/lib/components/editor/EditorLayout.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { LeftPanel, MainPanel, RightPanel, ThreeColumns } from '@design-system';
 import { ReactNode, use, useEffect, useState } from 'react';
 import { EditorContextProvider } from './EditorProvider';
 import {
@@ -8,7 +7,7 @@ import {
   getTranslationMetadataByUuid,
   Work,
 } from '@data-access';
-import { TranslationSkeleton } from '../shared/TranslationSkeleton';
+import { ThreeColumnRenderer, TranslationSkeleton } from '../shared';
 
 export const EditorLayout = ({
   left,
@@ -41,13 +40,7 @@ export const EditorLayout = ({
 
   return (
     <EditorContextProvider work={work}>
-      <div className="absolute fixed top-0 w-full bg-[url(/images/backgrounds/bg-reader.webp)] h-150 bg-[length:100%_auto] -z-10" />
-      <div className="absolute fixed top-0 w-full h-150 bg-gradient-to-b from-50% to-white -z-10" />
-      <ThreeColumns>
-        <LeftPanel>{left}</LeftPanel>
-        <MainPanel>{main}</MainPanel>
-        <RightPanel>{right}</RightPanel>
-      </ThreeColumns>
+      <ThreeColumnRenderer left={left} main={main} right={right} />
     </EditorContextProvider>
   );
 };
